WIP character skill progression: - Fixing an item increases the skill required to fix the item. - Fabricating something increases the skills required to fabricate the item. - Repairing a wall increases the mechanic skill.

+++ Barotrauma/BarotraumaShared/Source/Characters/CharacterNetworking.cs
@@ -584,7 +584,7 @@ namespace Barotrauma
                 foreach (Skill skill in info.Job.Skills)
                 {
                     msg.Write(skill.Name);
-                    msg.WriteRangedInteger(0, 100, MathHelper.Clamp(skill.Level, 0, 100));
+                    msg.WriteRangedInteger(0, 100, (int)MathHelper.Clamp(skill.Level, 0, 100));
                 }
             }
             else
